Use MathJax from Cloudflare CDN, remove async.
The MathJax CDN is retired. That means that demos like the simple forward pass will stop working in the future. I've pointed MathJax to a copy from the Cloudflare CDN.

In addition, the `async` attribute on the script means that the page actually has a race condition, where it will show up depending on whether the required JavaScript is already in cache.